+++ src/Http/ClientBuilder.php
@@ -17,7 +17,7 @@ use Gitlab\HttpClient\Builder;
 use GrahamCampbell\CachePlugin\CachePlugin;
 use Http\Client\Common\HttpMethodsClient;
 use Http\Client\Common\Plugin\Cache\Generator\CacheKeyGenerator;
-use Http\Client\Common\PluginClient;
+use Http\Client\Common\PluginClientFactory;
 use Psr\Cache\CacheItemPoolInterface;
 use ReflectionClass;
 
@@ -42,15 +42,15 @@ class ClientBuilder extends Builder
      */
     public function getHttpClient()
     {
-        if ($this->getProperty('httpClientModified')->getValue($this)) {
-            $this->getProperty('httpClientModified')->setValue($this, false);
-            $this->getProperty('pluginClient')->setValue($this, new HttpMethodsClient(
-                new PluginClient($this->getProperty('httpClient')->getValue($this), $this->getPlugins()),
-                $this->getProperty('requestFactory')->getValue($this)
+        if ($this->getPropertyValue('httpClientModified')) {
+            $this->setPropertyValue('httpClientModified', false);
+            $this->setPropertyValue('pluginClient', new HttpMethodsClient(
+                (new PluginClientFactory())->createClient($this->getPropertyValue('httpClient'), $this->getPlugins()),
+                $this->getPropertyValue('requestFactory')
             ));
         }
 
-        return $this->getProperty('pluginClient')->getValue($this);
+        return $this->getPropertyValue('pluginClient');
     }
 
     /**
@@ -60,7 +60,7 @@ class ClientBuilder extends Builder
      */
     protected function getPlugins()
     {
-        $plugins = $this->getProperty('plugins')->getValue($this);
+        $plugins = $this->getPropertyValue('plugins');
 
         if ($this->cachePlugin) {
             $plugins[] = $this->cachePlugin;
@@ -81,7 +81,7 @@ class ClientBuilder extends Builder
     {
         $this->setCachePlugin($cachePool, $config['generator'] ?? null, $config['lifetime'] ?? null);
 
-        $this->getProperty('httpClientModified')->setValue($this, true);
+        $this->setPropertyValue('httpClientModified', true);
     }
 
     /**
@@ -95,9 +95,34 @@ class ClientBuilder extends Builder
      */
     protected function setCachePlugin(CacheItemPoolInterface $cachePool, CacheKeyGenerator $generator = null, int $lifetime = null)
     {
-        $stream = $this->getProperty('streamFactory')->getValue($this);
+        $stream = $this->getPropertyValue('streamFactory');
 
         $this->cachePlugin = new CachePlugin($cachePool, $stream, $generator, $lifetime);
+    }
+
+    /**
+     * Get the value of the given private property on the builder.
+     *
+     * @param string $name
+     *
+     * @return mixed
+     */
+    protected function getPropertyValue(string $name)
+    {
+        return static::getProperty($name)->getValue($this);
+    }
+
+    /**
+     * Set the value of the given private property on the builder.
+     *
+     * @param string $name
+     * @param mixed  $value
+     *
+     * @return void
+     */
+    protected function setPropertyValue(string $name, $value)
+    {
+        static::getProperty($name)->setValue($this, $value);
     }
 
     /**
@@ -107,7 +132,7 @@ class ClientBuilder extends Builder
      *
      * @return \ReflectionProperty
      */
-    protected function getProperty(string $name)
+    protected static function getProperty(string $name)
     {
         $prop = (new ReflectionClass(Builder::class))->getProperty($name);