+++ CHANGELOG.md
@@ -2,6 +2,16 @@ CHANGE LOG
 ==========
 
 
-## V1.0 (Upcoming)
+## V1.2 (25/03/2018)
+
+* Support only m4tthumphrey/php-gitlab-api 9.8
+
+
+## V1.1 (24/03/2018)
+
+* Use PluginClientFactory in the builder
+
+
+## V1.0 (18/03/2018)
 
 * Initial release